Move keynames to header file
